indexページに実装を移植
